refactor(runtime): extract prompt execution module (#1280)
